feat: 쿠폰 서비스 API 연동 (쿠폰 검증 및 사용/취소 처리)

+++ order-service/src/main/java/com/palja/order_service/application/service/TimeDealService.java
@@ -10,8 +10,8 @@ public interface TimeDealService {
     TimeDealRes getTimeDeal(UUID timeDealId);
 
     // 타임딜 재고 차감
-    void deductTimeDealStock(UUID timeDealId, int quantity);
+    void deductTimeDealStock(UUID timeDealId, Long quantity);
 
     // 타임딜 재고 복구
-    void restoreTimeDealStock(UUID timeDealId, int quantity);
+    void restoreTimeDealStock(UUID timeDealId, Long quantity);
 }
